registerallocator: Elaborate module docstring.
This started as just type-fixing, but a chance was taken to extend the
description, to make it more suitable for people not familiar with the
topic of Register Allocation. This docstring also appears to serve
double role: it's both an intro into register allocation, and module
docstring per se (which is a separate issue which might need be addressing
eventually). To make it better intro, a descriptor was generalized somewhat,
to not give impression that there's only one right way to do something
(many points are heuristics, and different disciplines exist; indeed,
hopefully PPCI will support different register allocations and will
allow user to choose easily among them).